Moved scripts inside src
