+++ package.json
@@ -31,8 +31,8 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "add-puzzles": "ts-node --compiler-options '{\"module\": \"commonjs\"}' scripts/addPuzzles.ts",
-    "update-data": "ts-node --compiler-options '{\"module\": \"commonjs\"}' scripts/updateData.ts"
+    "add-puzzles": "ts-node --compiler-options '{\\\"module\\\": \\\"commonjs\\\"}' src/scripts/addPuzzles.ts",
+    "update-data": "ts-node --compiler-options '{\\\"module\\\": \\\"commonjs\\\"}' src/scripts/updateData.ts"
   },
   "eslintConfig": {
     "extends": "react-app"
