+++ src/scripts/analyticsDatabase.ts
